Change over to go querystring
Also merge everything into one file per type to clean up codebase
