+++ go.mod
@@ -2,4 +2,9 @@ module github.com/mellena1/mbta-v3-go
 
 go 1.12
 
-require github.com/michele/go.jsonapi v0.0.0-20190314113132-fa5f73018bc4
+require (
+	github.com/google/go-querystring v1.0.0
+	github.com/google/jsonapi v0.0.0-00010101000000-000000000000
+)
+
+replace github.com/google/jsonapi => github.com/michele/go.jsonapi v0.0.0-20190314113132-fa5f73018bc4
