+++ mbta/stops_test.go
@@ -28,7 +28,7 @@ func Test_GetStop(t *testing.T) {
 	mbtaClient := NewClient(ClientConfig{BaseURL: server.URL})
 	mbtaClient.client = server.Client()
 
-	actual, _, err := mbtaClient.Stops.GetStop(id, GetStopRequestConfig{})
+	actual, _, err := mbtaClient.Stops.GetStop(id, &GetStopRequestConfig{})
 	ok(t, err)
 	equals(t, expected, actual)
 }
@@ -68,7 +68,7 @@ func Test_GetAllStops(t *testing.T) {
 	mbtaClient := NewClient(ClientConfig{BaseURL: server.URL})
 	mbtaClient.client = server.Client()
 
-	actual, _, err := mbtaClient.Stops.GetAllStops(GetAllStopsRequestConfig{})
+	actual, _, err := mbtaClient.Stops.GetAllStops(&GetAllStopsRequestConfig{})
 	ok(t, err)
 	equals(t, expected, actual)
 }
